+++ CriticalMapsKit/Sources/AppFeature/AppView.swift
@@ -9,18 +9,19 @@ import SwiftUI
 
 /// The apps main view
 public struct AppView: View {
-  let store: Store<AppFeature.State, AppFeature.Action>
-  @ObservedObject var viewStore: ViewStore<AppFeature.State, AppFeature.Action>
+  @State var showsInfoExpanded = false
+  
+  let store: StoreOf<AppFeature>
+  @ObservedObject var viewStore: ViewStoreOf<AppFeature>
 
   @Environment(\.accessibilityReduceTransparency) private var reduceTransparency
   @Environment(\.horizontalSizeClass) private var horizontalSizeClass
 
   @State private var showOfflineBanner = false
 
-  private let minHeight: CGFloat = 56
-  public init(store: Store<AppFeature.State, AppFeature.Action>) {
+  public init(store: StoreOf<AppFeature>) {
     self.store = store
-    viewStore = ViewStore(store)
+    viewStore = ViewStore(store, observe: { $0 })
   }
 
   private var contextMenuTitle: String {
@@ -41,30 +42,46 @@ public struct AppView: View {
       )
       .edgesIgnoringSafeArea(.vertical)
 
-      VStack(alignment: .leading) {
-        if viewStore.state.mapFeatureState.isNextRideBannerVisible {
-          nextRideBanner
-            .contextMenu {
-              Button(
-                action: { viewStore.send(.set(\.$bottomSheetPosition, .relative(0.4))) },
-                label: { Label(contextMenuTitle, systemImage: "list.bullet") }
-              )
+      HStack {
+        VStack(alignment: .leading) {
+          if viewStore.mapFeatureState.isNextRideBannerVisible, viewStore.settingsState.rideEventSettings.isEnabled {
+            nextRideBanner()
+              .contextMenu {
+                Button(
+                  action: { viewStore.send(.set(\.$bottomSheetPosition, .relative(0.4))) },
+                  label: { Label(contextMenuTitle, systemImage: "list.bullet") }
+                )
+              }
+          }
+          
+          if viewStore.settingsState.infoViewEnabled {
+            ZStack(alignment: .center) {
+              Blur()
+                .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
+                .frame(width: showsInfoExpanded ? 120 : 50, height: showsInfoExpanded ? 230 : 50)
+                .accessibleAnimation(.cmSpring.speed(1.5), value: showsInfoExpanded)
+              
+              infoContent()
             }
+            .padding(.bottom, .grid(2))
+          }
+          
+          if viewStore.hasOfflineError {
+            offlineBanner()
+              .clipShape(Circle())
+              .opacity(showOfflineBanner ? 1 : 0)
+              .accessibleAnimation(.easeInOut(duration: 0.2), value: showOfflineBanner)
+          }
         }
+        .padding(.top, .grid(1))
 
-        if !viewStore.connectionObserverState.isNetworkAvailable || viewStore.hasOfflineError {
-          offlineBanner
-            .clipShape(Circle())
-            .opacity(showOfflineBanner ? 1 : 0)
-            .accessibleAnimation(.easeOut, value: showOfflineBanner)
-        }
+        Spacer()
       }
-      .padding(.top, .grid(2))
       .padding(.horizontal)
 
       VStack {
         Spacer()
-
+        
         AppNavigationView(store: store)
           .accessibilitySortPriority(1)
           .padding(.horizontal)
@@ -87,14 +104,69 @@ public struct AppView: View {
     .showDragIndicator(true)
     .enableSwipeToDismiss()
     .onDismiss { viewStore.send(.set(\.$bottomSheetPosition, .hidden)) }
-    .alert(store.scope(state: \.alert), dismiss: .dismissAlert)
+    .alert(store.scope(state: \.alert, action: { $0 }), dismiss: .dismissAlert)
     .onAppear { viewStore.send(.onAppear) }
     .onDisappear { viewStore.send(.onDisappear) }
-    .onChange(of: viewStore.connectionObserverState.isNetworkAvailable) { newValue in
-      self.showOfflineBanner = !newValue
-    }
   }
 
+  @ViewBuilder
+  func infoContent() -> some View {
+    if showsInfoExpanded {
+      VStack {
+        Text("Info")
+          .foregroundColor(Color(.textPrimary))
+          .font(.titleTwo)
+        
+        DataTile("Next update") {
+          CircularProgressView(progress: viewStore.timerProgress)
+            .frame(width: 44, height: 44)
+            .overlay(alignment: .center) {
+              if viewStore.isRequestingRiderLocations {
+                ProgressView()
+              } else {
+                Text(verbatim: viewStore.timerValue)
+                  .foregroundColor(Color(.textPrimary))
+                  .font(.system(size: 14).bold())
+                  .monospacedDigit()
+              }
+            }
+            .padding(.top, .grid(1))
+        }
+  
+        DataTile("Riders") {
+          HStack {
+            Text(viewStore.ridersCount)
+              .font(.pageTitle)
+          }
+          .foregroundColor(Color(.textPrimary))
+        }
+      }
+      .transition(
+        .asymmetric(
+          insertion: .opacity.combined(with: .scale(scale: 1, anchor: .topLeading)).animation(.easeIn(duration: 0.2)),
+          removal: .opacity.combined(with: .scale(scale: 0, anchor: .topLeading)).animation(.easeIn(duration: 0.12))
+        )
+      )
+      .contentShape(Rectangle())
+      .onTapGesture {
+        withAnimation { showsInfoExpanded = false }
+      }
+    } else {
+      Button(action: { withAnimation { showsInfoExpanded = true } }) {
+        Image(systemName: "info.circle")
+          .resizable()
+          .frame(width: 30, height: 30)
+          .transition(
+            .asymmetric(
+              insertion: .opacity.combined(with: .scale(scale: 0, anchor: .bottomLeading)).animation(.easeIn(duration: 0.1)),
+              removal: .opacity.animation(.easeIn(duration: 0.1))
+            )
+          )
+      }
+    }
+  }
+  
+  @ViewBuilder
   func bottomSheetContentView() -> some View {
     VStack {
       List(viewStore.nextRideState.rideEvents, id: \.id) { ride in
@@ -145,7 +217,8 @@ public struct AppView: View {
     }
   }
 
-  var offlineBanner: some View {
+  @ViewBuilder
+  func offlineBanner() -> some View {
     Image(systemName: "wifi.slash")
       .foregroundColor(
         reduceTransparency
@@ -158,7 +231,7 @@ public struct AppView: View {
         Group {
           if reduceTransparency {
             RoundedRectangle(
-              cornerRadius: 12,
+              cornerRadius: 8,
               style: .circular
             )
             .fill(reduceTransparency
@@ -172,14 +245,18 @@ public struct AppView: View {
       )
   }
 
-  var nextRideBanner: some View {
+  @ViewBuilder
+  func nextRideBanner() -> some View {
     MapOverlayView(
-      store: store.actionless.scope(state: {
-        MapOverlayView.ViewState(
-          isVisible: $0.mapFeatureState.isNextRideBannerVisible,
-          isExpanded: $0.mapFeatureState.isNextRideBannerExpanded
-        )
-      }),
+      store: store.scope(
+        state: {
+          MapOverlayView.ViewState(
+            isVisible: $0.mapFeatureState.isNextRideBannerVisible,
+            isExpanded: $0.mapFeatureState.isNextRideBannerExpanded
+          )
+        },
+        action: { $0 }
+      ).actionless,
       action: { viewStore.send(.map(.focusNextRide(viewStore.nextRideState.nextRide?.coordinate))) },
       content: {
         VStack(alignment: .leading, spacing: .grid(1)) {