+++ CriticalMapsKit/Sources/MapFeature/MapFeatureView.swift
@@ -8,17 +8,14 @@ import SwiftUI
 
 public struct MapFeatureView: View {
   @Environment(\.accessibilityReduceTransparency) var reduceTransparency
-
-  public typealias State = MapFeature.State
-  public typealias Action = MapFeature.Action
-
-  public init(store: Store<State, Action>) {
+  
+  public init(store: StoreOf<MapFeature>) {
     self.store = store
-    viewStore = ViewStore(store)
+    viewStore = ViewStore(store, observe: { $0 })
   }
 
-  let store: Store<State, Action>
-  @ObservedObject var viewStore: ViewStore<State, Action>
+  let store: StoreOf<MapFeature>
+  @ObservedObject var viewStore: ViewStoreOf<MapFeature>
 
   public var body: some View {
     ZStack(alignment: .topLeading) {
@@ -27,6 +24,7 @@ public struct MapFeatureView: View {
         userTrackingMode: viewStore.binding(\.$userTrackingMode),
         nextRide: viewStore.nextRide,
         rideEvents: viewStore.rideEvents,
+        annotationsCount: viewStore.binding(\.$visibleRidersCount),
         centerRegion: viewStore.binding(\.$centerRegion),
         centerEventRegion: viewStore.binding(\.$eventCenter),
         mapMenuShareEventHandler: {